+++ lib/src/data/google_places_api.dart
@@ -2,13 +2,31 @@ import 'dart:convert';
 
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:google_places_autocomplete_text_field/src/model/place_details.dart';
-import 'package:google_places_autocomplete_text_field/src/model/prediction.dart';
+import 'package:google_places_autocomplete_text_field/google_places_autocomplete_text_field.dart';
+
+/// {@template places_api}
+/// Interface for Places Autocomplete requests.
+/// {@endtemplate}
+abstract class PlacesApi {
+  /// Fetches suggestions for the given input with the provided
+  /// [GoogleApiConfig].
+  Future<PlacesAutocompleteResponse?> getSuggestionsForInput({
+    required String input,
+    required GoogleApiConfig config,
+  });
+
+  /// Fetches the place details for the given [Prediction] with the provided
+  /// [GoogleApiConfig].
+  Future<Prediction?> fetchCoordinatesForPrediction({
+    required Prediction prediction,
+    required GoogleApiConfig config,
+  });
+}
 
 /// {@template google_places_api}
 /// Interface for Google Places API.
 /// {@endtemplate}
-class GooglePlacesApi {
+class GooglePlacesApi implements PlacesApi {
   /// {@macro google_places_api}
   GooglePlacesApi() : _dio = Dio();
 
@@ -18,52 +36,62 @@ class GooglePlacesApi {
   /// The API URL for the Google Places API.
   final _apiUrl = 'https://places.googleapis.com/v1/places';
 
-  /// Fetches suggestions for the given [input] with the provided
-  /// [googleAPIKey].
+  @override
   Future<PlacesAutocompleteResponse?> getSuggestionsForInput({
     required String input,
-    required String googleAPIKey,
-    List<String> countries = const [],
-    String? sessionToken,
-    String proxyUrl = '',
-    String? languageCode,
+    required GoogleApiConfig config,
   }) async {
-    final prefix = proxyUrl;
+    final prefix = config.proxyURL ?? '';
 
-    String url = "$prefix$_apiUrl:autocomplete";
+    String url = '$prefix$_apiUrl:autocomplete';
 
-    Map<String, dynamic> requestBody = {
-      "input": input,
-    };
+    Map<String, dynamic> requestBody = {'input': input};
 
-    if (countries.isNotEmpty) {
-      requestBody["includedRegionCodes"] = countries;
-      requestBody["languageCode"] = languageCode;
+    if (config.countries.isNotEmpty) {
+      requestBody['includedRegionCodes'] = config.countries;
+      requestBody['languageCode'] = config.languageCode;
     }
-    if (sessionToken != null) {
-      requestBody["sessionToken"] = sessionToken;
+    if (config.sessionToken != null) {
+      requestBody['sessionToken'] = config.sessionToken;
     }
+
+    // Only one of locationRestriction or locationBias can be provided,
+    // so we prefer restriction over bias, as it is more exclusive.
+    if (config.locationRestriction != null) {
+      requestBody['locationRestriction'] = config.locationRestriction?.toJson();
+    } else if (config.locationBias != null) {
+      requestBody['locationBias'] = config.locationBias?.toJson();
+    }
+
+    if (config.placeTypeRestriction != null) {
+      requestBody['includedPrimaryTypes'] =
+          config.placeTypeRestriction!.toJson();
+    }
+
     Options options = Options(
-      headers: {
-        "X-Goog-Api-Key": googleAPIKey,
-        "X-Goog-FieldMask": "*",
-      },
+      headers: {'X-Goog-Api-Key': config.apiKey, 'X-Goog-FieldMask': '*'},
     );
 
     try {
-      final response =
-          await _dio.post(url, options: options, data: jsonEncode(requestBody));
-      final subscriptionResponse =
-          PlacesAutocompleteResponse.fromJson(response.data);
+      final response = await _dio.post(
+        url,
+        options: options,
+        data: jsonEncode(requestBody),
+      );
+      final subscriptionResponse = PlacesAutocompleteResponse.fromJson(
+        response.data,
+      );
       return subscriptionResponse;
     } on DioException catch (e) {
       if (e.response != null) {
         debugPrint(
-            'GooglePlacesApi.getSuggestionsForInput: DioException [${e.type}]: ${e.message}');
+          'GooglePlacesApi.getSuggestionsForInput: DioException [${e.type}]: ${e.message}',
+        );
         debugPrint('Response data: ${e.response?.data}');
       } else {
         debugPrint(
-            'GooglePlacesApi.getSuggestionsForInput: DioException [${e.type}]: ${e.message}');
+          'GooglePlacesApi.getSuggestionsForInput: DioException [${e.type}]: ${e.message}',
+        );
       }
       return null;
     } catch (e) {
@@ -72,21 +100,19 @@ class GooglePlacesApi {
     }
   }
 
-  /// Fetches the place details for the given [prediction] with the provided
-  /// [googleAPIKey].
+  @override
   Future<Prediction?> fetchCoordinatesForPrediction({
     required Prediction prediction,
-    required String googleAPIKey,
-    String proxyUrl = '',
-    String? sessionToken,
+    required GoogleApiConfig config,
   }) async {
     try {
-      final prefix = proxyUrl;
+      final prefix = config.proxyURL ?? '';
 
       String url =
-          "$prefix$_apiUrl/${prediction.placeId}?fields=*&key=$googleAPIKey";
+          '$prefix$_apiUrl/${prediction.placeId}?fields=*&key=${config.apiKey}';
+      final sessionToken = config.sessionToken;
       if (sessionToken != null) {
-        url += "&sessionToken=$sessionToken";
+        url += '&sessionToken=$sessionToken';
       }
       final response = await _dio.get(url);
 
@@ -99,16 +125,19 @@ class GooglePlacesApi {
     } on DioException catch (e) {
       if (e.response != null) {
         debugPrint(
-            'GooglePlacesApi.fetchCoordinatesForPrediction: DioException [${e.type}]: ${e.message}');
+          'GooglePlacesApi.fetchCoordinatesForPrediction: DioException [${e.type}]: ${e.message}',
+        );
         debugPrint('Response data: ${e.response?.data}');
       } else {
         debugPrint(
-            'GooglePlacesApi.fetchCoordinatesForPrediction: DioException [${e.type}]: ${e.message}');
+          'GooglePlacesApi.fetchCoordinatesForPrediction: DioException [${e.type}]: ${e.message}',
+        );
       }
       return null;
     } catch (e) {
       debugPrint(
-          'GooglePlacesApi.fetchCoordinatesForPrediction: ${e.toString()}');
+        'GooglePlacesApi.fetchCoordinatesForPrediction: ${e.toString()}',
+      );
       return null;
     }
   }