+++ lib/src/model/model_barrel.dart
@@ -1,2 +1,3 @@
+export 'google_api_config.dart';
 export 'place_details.dart';
 export 'prediction.dart';